+++ tests/testthat/test-dataset-fitting.R
@@ -1,7 +1,7 @@
 context("dataset-fitting")
 
 
-library(keras)
+library(keras3)
 
 test_succeeds("can fit a dataset using keras fit", {
   d  <-
@@ -9,25 +9,12 @@ test_succeeds("can fit a dataset using keras fit", {
     dataset_batch(2) %>%
     dataset_shuffle(1024)
 
-  model <- keras_model_sequential() %>%
-    layer_dense(units = 1, input_shape = 2)
+  skip_if(tf_version() < "2.16")
+  model <-
+    keras_model_sequential(input_shape = 2) %>%
+    layer_dense(units = 1)
 
   model %>% compile(optimizer = "adam", loss = "mse")
   model %>% fit(d, steps_per_epoch = 5, epochs = 1)
 })
 
-test_succeeds("can fit a dataset using keras fit_generator", {
-  d  <-
-    tensor_slices_dataset(list(matrix(1:20, ncol = 2), matrix(1:10, ncol = 1))) %>%
-    dataset_batch(2) %>%
-    dataset_shuffle(1024)
-
-
-  model <- keras_model_sequential() %>%
-    layer_dense(units = 1, input_shape = 2)
-
-  model %>% compile(optimizer = "adam", loss = "mse")
-  expect_warning( {
-    model %>% fit_generator(d, steps_per_epoch = 5, epochs = 1)
-  }, "deprecated")
-})